Check for name conflicts between stageusers and users
Check if the username conflicts with an existing user. The check
is not perfect. A user might be created before the stage user is
activated. The code also suffers from a race condition. It's as
good as it can get without a better API, though.

closes #10

see https://fedorahosted.org/freeipa/ticket/5186